+++ src/components/forms/FormTextInput.ts
@@ -1,0 +1,20 @@
+//
+// Imports
+//
+
+import { DE, InputElementAttributes } from "@donutteam/document-builder";
+
+//
+// Exports
+//
+
+export function FormTextInput(attributes : InputElementAttributes) : DE
+{
+	return new DE("input",
+		{
+			class: "component-form-text-input",
+			type: "text",
+
+			...attributes,
+		});
+}